各種アニメを別ファイルに切り出した
