+++ src/js/td-scenes/battle/animation/game-state/battle/attack/shin-braver/feint.ts
@@ -1,0 +1,18 @@
+import { Feint } from "gbraver-burst-core";
+
+import { Animate } from "../../../../../../../animation/animate";
+import { delay, empty } from "../../../../../../../animation/delay";
+import { ShinBraverBattle } from "./shin-braver-battle";
+
+/**
+ * フェイント
+ * @param param パラメータ
+ * @returns アニメーション
+ */
+export function feint(param: ShinBraverBattle<Feint>): Animate {
+  if (!param.result.isDefenderMoved) {
+    return empty();
+  }
+
+  return param.defenderSprite.avoid().chain(delay(500));
+}
